% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_treearchi_group_mlr.R
\name{run_treearchi_group_mlr}
\alias{run_treearchi_group_mlr}
\title{Run group-specific TreeArchi MLR workflow}
\usage{
run_treearchi_group_mlr(
  data_path,
  out_base,
  response_var,
  group_var,
  predictor_vars,
  group_reference = NULL,
  random_effect_var = "genus",
  id_col = "tls_id",
  quadratic_vars = NULL,
  interaction_pairs = NULL,
  analysis_tag = "GROUP_MLR",
  run_diagnostics = TRUE,
  make_coef_plot = TRUE,
  make_group_lineplots = TRUE
)
}
\description{
Run group-specific TreeArchi MLR workflow
}
