% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_treearchi_mlr}
\alias{run_treearchi_mlr}
\title{Run the main TreeArchi MLR workflow}
\usage{
run_treearchi_mlr(
  data_path,
  out_base,
  response_var,
  forced_var,
  predictor_vars,
  random_effect_var = "genus",
  id_col = "tls_id",
  analysis_tag = "TREEARCHI_MLR",
  make_coef_plot = TRUE,
  make_diagnostics = TRUE,
  make_local_effect_lineplots = TRUE,
  label_n = 5
)
}
\description{
Run the main TreeArchi MLR workflow
}
